Draine scattering and code for reading BPASS data
Added a simple bit of code to read BPASS datasets.

Support for Draine scattering function added.  Still need to find the limits where Draine is innacurate, as is we cannot approach g=1 or -1 with Draine, since the accept/reject method will take an arbitraty long time to finish with those parameters.  Those are best handled by HG, but nearby values of g should probably be handled by HG as well.